fix: cast TIOCSWINSZ/TIOCGWINSZ to c_ulong for macOS ioctl compatibility
On macOS, the ioctl function expects c_ulong (u64) for the request
parameter, but TIOCSWINSZ and TIOCGWINSZ are defined as u32. This
is the same issue that was fixed for TIOCSCTTY.

+++ terminal/terminal-pty/src/pty.rs
@@ -80,7 +80,8 @@ impl Pty {
         let ws = size.to_winsize();
         let fd = self.master.as_raw_fd();
 
-        let result = unsafe { libc::ioctl(fd, libc::TIOCSWINSZ, &ws) };
+        // Note: On macOS, ioctl expects c_ulong for request parameter
+        let result = unsafe { libc::ioctl(fd, libc::TIOCSWINSZ as libc::c_ulong, &ws) };
 
         if result == -1 {
             Err(Error::WindowSize(io::Error::last_os_error().to_string()))
@@ -94,7 +95,8 @@ impl Pty {
         let mut ws: libc::winsize = unsafe { std::mem::zeroed() };
         let fd = self.master.as_raw_fd();
 
-        let result = unsafe { libc::ioctl(fd, libc::TIOCGWINSZ, &mut ws) };
+        // Note: On macOS, ioctl expects c_ulong for request parameter
+        let result = unsafe { libc::ioctl(fd, libc::TIOCGWINSZ as libc::c_ulong, &mut ws) };
 
         if result == -1 {
             Err(Error::WindowSize(io::Error::last_os_error().to_string()))
